[refactor]: Extract macro utils into `iroha_macro_utils`
