+++ config/base/derive/Cargo.toml
@@ -13,3 +13,5 @@ syn = { version = "1.0.107", default-features = false, features = ["derive", "pa
 quote = "1.0.23"
 proc-macro2 = "1.0.49"
 proc-macro-error = "1.0.4"
+
+iroha_macro_utils = { version = "=2.0.0-pre-rc.13", path = "../../../macro/utils" }
